Update always-on-extended-events.md
According to my check on over 10 instances of SQL Server, there is a mistake in the following sentence: "The alwayson_health extended events session is created automatically when you create the availability group". This extended event is created even if you do not use availability group. It exists even in new clean installation when I selects only the Data Engine feature. The next sentence is the way it is behave "The Create Availability Group Wizard automatically starts the session".